mod can not import py mod from ansible_collection
The way anziballz are constructed currently, this form of
imports will fail:

from ansible_collections.mynamespace.myname.plugins.module_utils import some_mod_util

That results in the the recursive_finder attempting to load a python
module from a file called:
        ~/.ansible/content/ansible_collections/mynamespace/myname/plugins/module_utils.py

That load fails, and the module_common ansiballZ creater stops trying to
find python modules to include in the anziball, resulting in the task
module execution failing with import errors when it doesn't find
a 'ansible_collections' top level.

So for now, don't do that in this module.